Change the integration grid size and type of an Echoview acoustic variable

This function sets the grid separation and depth reference line for an acoustic variable using COM scripting.

EVChangeVariableGrid(EVFile, acoVarName, timeDistanceGridType = NULL,
  depthGridType = NULL, timeDistanceGridDistance = NULL,
  depthGridDistance = NULL, EVLineName = NULL, showenum = FALSE)

Arguments

EVFile
An Echoview file COM object

acoVarName
(character) Echoview acoustic variable name

timeDistanceGridType
(numeric) specifying the along track grid type. 0 = no grid, 1 = time (minutes), 2 = GPS distance (NMi), 3 = Vessel Log Distance (NMi), 4 = Pings, 5 = GPS distance (m), 6 = Vessel Log Distance (m).

depthGridType
(numeric) 0 = no grid, 1 = depth grid, 2 = use reference line.

timeDistanceGridDistance
(numeric) vertical grid line spacing. Not needed if verticalType = 0.

depthGridDistance
(numeric) horizontal grid line spacing. Ignored if horizontalType = 0.

EVLineName
(character) an Echoview line object name. Not needed if depthGridType != 2.

showenum
= FALSE (boolean) if TRUE Echoview grid enum values are returned to the workspace. All other arguments are ignored.

Value

a list object with three elements. $predataRangeSettings: a vector of pre-function call settings; $postDataRangeSettings a vector of post-function call settings, and $msg: message for processing log.

Details

depthGridType is specified in the Echoview COM enum values called 'EDepthRangeGridMode'. TimeDistanceGridMode is specified in the Echoview COM enum values called ETimeDistanceGridMode

Note

This function was completely rewritten Sept. 2017 which resulted in changes to function argument names.

References

http://support.echoview.com/WebHelp/Echoview.htm/

See also

Examples

# NOT RUN {
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

EVChangeVariableGrid <- function (EVFile,
acoVarName="38 seabed and surface excluded",
timeDistanceGridType=5, depthGridType=2,
timeDistanceGridDistance = 100, depthGridDistance = 10,
EVLineName = "Fixed depth 250 m")
#remove grid:
EVChangeVariableGrid <- function (EVFile,
acoVarName="38 seabed and surface excluded",
timeDistanceGridType=0, depthGridType=0)
#only show Echoview grid enum values
EVChangeVariableGrid(showenum = TRUE)
# }