Create a new Echoview (.EV) file and adds raw data files to it

This function creates a new Echoview (.EV) file and adds raw data files to it via COM scripting. Works well when populating an existing Echoview template file with raw data files. The newly created Echoview file will remain open in Echoview and can be accessed via the $EVFile objected returned by a successful call of this function.

Usage

EVCreateNew(
  EVAppObj,
  templateFn = NULL,
  EVFileName = NULL,
  filesetName = "Fileset1",
  dataFiles = NULL,
  CloseOnSave = FALSE
)

Arguments

EVAppObj

An EV application COM object arising from the call COMCreate('EchoviewCom.EvApplication')

templateFn

= NULL Full path and filename of template file if used.

EVFileName

Full path and filename of Echoview (.EV) file to be created.

filesetName

Echoview fileset name

dataFiles

vector of full path and name for each data file.

CloseOnSave

= TRUE close the EV file in EVFileName once saved.

Value

a list object with two elements. $EVFile: EVFile COM object for the newly created Echoview file, and $msg: message for processing log.

Details

For the example code to run, the example data must be downloaded from. NB the example code assumes the data and directory structure of the example data has been maintained.

References

http://support.echoview.com/WebHelp/Echoview.htm/

See also

Examples

if (FALSE) {
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
pathAndFn=list.files("~~/KAOS/raw/", full.names=TRUE)
#remove any evi type files
eviLoc=grep('.evi',pathAndFn)
if(length(eviLoc)>0) (pathAndFn=pathAndFn[-eviLoc])
EVCreateNew(EVAppObj=EVAppObj,
                 templateFn="~~/KAOS/KAOStemplate.EV",
                 EVFileName='~~/KAOS/kaos.ev',
                 filesetName="038-120-200",
                 dataFiles=pathAndFn, 
                 CloseOnSave = TRUE)
}