Export Wideband Single Targets Frequency Response

This function exports the broadband single targets frequency response

EVExportWBST_FR(EVFile = EVFile, STVar, outfn, AverageResults = 0,
  MinMax = 0, WindowSize = 1, WindowUnit = 2, EVVar)

Arguments

EVFile
An Echoview file COM object

STVar
An Wideband Single Target Detection Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)

outfn
FIlename for the TS export

MinMax
Include Min and Max

WindowSize
Set the WIndow Size

WindowUnit
Unit of the window 0 = Meters, 1 = Pulse Lengths, 2 = Samples

EVVar
Specifies a TS pulse compressed wideband variable for frequency response calculations.

References

http://support.echoview.com/WebHelp/Echoview.htm/

Examples

#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()

#> 2018-11-01 10:21:43: Opening Echoview

#> Warning: Couldn't get clsid from the string

#> Error: Invalid class string

#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile

#> 2018-11-01 10:21:43 : Creating new EV file

#> Error in EVNewFile(EVAppObj = EVAppObj, templateFn = templateFn): object 'echoview' not found

#Set the First and second Operand
EVVar <- "Fileset 1: TS pulse compressed wideband pings T1"
Operand2 <- "Fileset 1: angular position pulse compressed wideband pings T1"
#Create a wideband single target detection varable
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2)

#> Error in EVAcoVarNameFinder(EVFile, acoVarName = EVVar): object 'EVFile' not found

# Change TS detection threshold
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2,TsThreshold=-90)

#> Error in EVAcoVarNameFinder(EVFile, acoVarName = EVVar): object 'EVFile' not found

#Export the single target frequency response
EVExportWBST_FR(ST,"WB_ST_FR.csv",0,0,1,2,EVVar)

#> Error in EVAcoVarNameFinder(EVFile, acoVarName = STVar): object 'ST' not found