Schools Detection in Echoview

This function performs schools detection in Echoview using COM scripting.

Usage

EVSchoolsDetect(
  EVApp = NULL,
  EVFile = NULL,
  acoVarName,
  outputRegionClassName,
  deleteExistingRegions,
  distanceMode,
  maximumHorizontalLink,
  maximumVerticalLink,
  minimumCandidateHeight,
  minimumCandidateLength,
  minimumSchoolHeight,
  minimumSchoolLength,
  dataThreshold
)

Arguments

EVApp

=NULL An Echoview Application COM object

EVFile

=NULL An Echoview file COM object

acoVarName

A string containing the name of the acoustic variable to perform the analysis on

outputRegionClassName

A string containing the name of the output region

deleteExistingRegions

Logical TRUE or FALSE

distanceMode

for schools detection (see Echoview help).

maximumHorizontalLink

The maximum horizontal link in meters

maximumVerticalLink

The maximum vertical link in meters

minimumCandidateHeight

The minimum candidate height in meters

minimumCandidateLength

the minimum candidate length in meters

minimumSchoolHeight

The minimum school height in meters

minimumSchoolLength

The minimum school length in meters

dataThreshold

minimum integration threshold (units: dB re 1m^-1)

Value

a list object with four elements. $nbrOfDetectedschools, $thresholdData, $schoolsSettingsData, and $msg: message for processing log

References

http://support.echoview.com/WebHelp/Echoview.htm/

See also

Examples

if (FALSE) {
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVSchoolsDetect(EVFile = EVFile,
               acoVarName='120 7x7 convolution',
               outputRegionClassName = 'aggregations',
               deleteExistingRegions = TRUE,
               distanceMode = "GPS distance",
               maximumHorizontalLink = 15, #m
               maximumVerticalLink = 5,#m
               minimumCandidateHeight = 1, #m
               minimumCandidateLength = 10, #m
               minimumSchoolHeight = 2, #m
               minimumSchoolLength = 15, #m
               dataThreshold = -80)
}