Change the single target settings for Echoview single target virtual variable

Change the single target settings for Echoview virtual variable

EVsingleTargetPars(EVFile, acoVarName, ExcludeAboveLine = NULL,
  ExcludeBelowLine = NULL, MinimumPulseLength = NULL,
  MaximumPulseLength = NULL, PLDL = NULL, TsThreshold = NULL,
  verbose = FALSE)

Arguments

EVFile
("COMIDispatch) An Echoview file COM object

acoVarName
(character) Name of Echoview single target virtual variable

ExcludeAboveLine
(character) Shallower exclusion line

ExcludeBelowLine
(character) Deeper exclusion line

MinimumPulseLength
(numeric) Minimum pulse length

MaximumPulseLength
(numeric) Maximum pulse length

PLDL
(numeric) Pulse length determination level

TsThreshold
(numeric) minimum target strength threshold (dB)

verbose
= FALSE (boolean)

Value

list $settings 2 column data frame of pre-change post-change parameter values, $msg= vector of messages

Details

This function will work with single target detection viritual variables

Note

THis function has only been tested using Simrad EK60 split-beam data