Write a map info file for import into Echoview.

This function writes a polygon MIF file for import into Echoview and will typically be used to export survey line transects from, for example, centreLawnOnPosition.

Usage

exportMIF(
  coords,
  pathAndFileName,
  pointNameExport = FALSE,
  pointNameScaleFactor = c(0.05, 0.03)
)

Arguments

coords

a set of coordinates as a two column (longitude and latitude) matrix.

pathAndFileName

character string MIF file export path and filename

pointNameExport

Boolean (default=FALSE) export the name of each point.

pointNameScaleFactor

=c(0.005,0.03) see details

Value

Nothing

Details

Exporting the name of each point uses the row name in the coords argument.

The pointNameScaleFactor argument is a two element numeric vector specifying the bounding box for each point name. The bounding box is calculated for point name i by specifying the lower left hand corner of the point name position, with the upper right hand corner specified as c(lon_i,lat_i)+diff(range(lon_1..n,lat_1..n))*pointNameScaleFactor where n is the total number of points i.e nrow(coords)

Examples

if (FALSE) {
coords=centreLawnOnPosition(centreLon=-33,centreLat=-55.5,lineLengthkm=100,lineSpacingkm=10,
startBearingdeg=30,numOfLines=50)
exportMIF(coords=coords,pathAndFileName='c:\\Users\\martin_cox\\Documents\\test4.mif')
}