Find Raw files based on Time and Date, providing folder containing all Raw files

This functions generates datetime object based on the raw data filenames and selects the required raw files according to a user defined start and end time

Usage

getRaws(
  dir,
  StartDate,
  StartTime,
  EndDate,
  EndTime,
  tzraw = "UTC",
  tzinp = "UTC",
  timebuffer = 0,
  fileSlope = FALSE
)

Arguments

dir

= path [character]

StartDate

[character] Date under the shape of YYYYMMDD (example: 20171127)

StartTime

[character] Time under the shape of HHMMSS (example: 085617)

EndDate

[character] Date under the shape of YYYYMMDD (example: 20171127)

EndTime

[character] Time under the shape of HHMMSS (example: 085617)

tzraw

="UTC"[character] timezone under which the raw files were recorded e.g.UTC or Australia/Brisbane or AEST

tzinp

="UTC"[character] timezone of the entered start and end points e.g.UTC or Australia/Brisbane or AEST

timebuffer

= 0 [numeric in seconds] add a buffer around the times for which files should be selected

fileSlope

= FALSE [boolean] If TRUE the raw file that partly falls within the start of the time interval.

Value

charater array with raw data filenames

Note

Dependencies: lubridate - Requires lubridate package for easy computation of time intervals

Examples

if (FALSE) {
StartDate = 20170901
StartTime = "080917"
EndDate = 20170901
EndTime = 172356
rawdir = "F:\\RawFolder\\"
getRaws(dir,StartDate, StartTime,EndDate, EndTime, tzraw="UTC", tzinp="UTC")
 }