Make a rectangular Echoview region (EVR) file.

This function creates a rectangular (2D) Echoview region file. The function can add regions to an existing EVR file.

makeRectRegion(dateStart, timeStart, dateStop, timeStop, minDepth = 0,
  maxDepth, outPathFn, regionClassName, regionType, append = FALSE)

Arguments

dateStart
date stamp for the start of the region (yyymmdd)

timeStart
time stamp for the start of the region (hhmmss)

dateStop
date stamp for the end of the region (yyymmdd)

timeStop
time stamp for the end of the region (hhmmss)

minDepth
=0 Minimum depth for the region

maxDepth
maximum depth for the region

outPathFn
output parth and filename.

regionClassName
region class name

regionType
region Type (integer)

append
= FALSE should be region be added to the EVR file given in outPathFn

Value

list: $msg messages from the function call $exportFLAG Boolean TRUE EVR file created or added. FALSE otherwise

Details

Region types supported by Echoview: 0 = bad (no data); 1 = analysis; "2" = marker; 3 = fishtracks, and 4 = bad (empty water)

References

http://support.echoview.com/WebHelp/Echoview.htm/

Examples

# NOT RUN {
makeRectRegion(dateStart='20170815',timeStart='001525',
dateStop='20170815',timeStop='001738',
minDepth=20,maxDepth=100,
outPathFn='~/test2.evr',
regionClassName='ctd',append=FALSE)
# }