Read CSV resulting from a call of EVExportUnderlying()

This function reads a CSV file created after a call of EVExportUnderlying(). The data can be Sv values or values from, for example, the Echoview formula operator.

readUnderlyingDataCSV(pathAndFn, samplePreFix = "sample",
  depthPreFix = "z", asList = FALSE, onlyHeader = FALSE,
  lastHeaderCol = 13)

Arguments

pathAndFn
The path and filename of the CSV fileonlyHeader

samplePreFix
='sample' (character)text to add to the sample column names

depthPreFix
= 'z' (character) text to add to the sample row names (depth)

asList
= FALSE (boolean) return the CSV as a list. See details.

onlyHeader
= FALSE (boolean) return only the header of the CSV, ignored if asList = TRUE. See details.

lastHeaderCol
= 13 The last column of the header information.

Value

see details

Details

The default return option is a data.frame of ping header and sample information. If asList=TRUE a list with $header and $samples is returned, If onlyHeader=TRUE only the headerinformation is returned.

References

http://support.echoview.com/WebHelp/Echoview.htm/

See also