Calculate the relative dimension of a school

This function calculates the relative school dimension of Diner (2001)

schoolRelativeDimension(Sv, L, P, Th, theta)

Arguments

Sv
Observed mean volume backscattering strength (Sv) of a school

L
Observed (uncorrected) school length, m

P
Mean depth of school

Th
Minimum Sv data threshold used to detect schools using the SHAPES algorithm

theta
Transducer 3 dB beam width, degrees

Value

Relative school dimension (Nb)

References

Diner, N. (2001). Correction on school geometry and density: approach based on acoustic image simulation. Aquatic Living Resources, 14(4), 211-222.

See also